+++ src/core/md_ring_allocator.h
@@ -15,13 +15,13 @@ extern "C" {
 // This is a small, efficient linear allocator which just pushes a pointer offset within a buffer and wraps around when full
 // For simplicity, all allocations > 2 bytes are aligned to 16 bytes
 typedef struct md_ring_allocator_t {
-    uint64_t pos;
-    uint64_t cap;
+    size_t   pos;
+    size_t   cap;
     uint64_t magic;
     void*    ptr;
 } md_ring_allocator_t;
 
-static inline void md_ring_allocator_init(md_ring_allocator_t* ring, void* backing_buffer, uint64_t buffer_capacity) {
+static inline void md_ring_allocator_init(md_ring_allocator_t* ring, void* backing_buffer, size_t buffer_capacity) {
     ASSERT(ring);
     ASSERT(backing_buffer);
     ASSERT(buffer_capacity > 0);
@@ -32,24 +32,24 @@ static inline void md_ring_allocator_init(md_ring_allocator_t* ring, void* backi
     ring->ptr = backing_buffer;
 }
 
-static inline void* md_ring_allocator_push_aligned(md_ring_allocator_t* ring, uint64_t size, uint64_t align) {
+static inline void* md_ring_allocator_push_aligned(md_ring_allocator_t* ring, size_t size, size_t align) {
     ASSERT(ring && ring->magic == MD_RING_ALLOCATOR_MAGIC);
     ASSERT(IS_POW2(align));
 
-    uint64_t pos_addr = (uint64_t)ring->ptr + ring->pos;
-    uint64_t alignment_size = ALIGN_TO(pos_addr, align) - pos_addr;
-    uint64_t pos = (ring->pos + alignment_size + size <= ring->cap) ? ring->pos + alignment_size : ALIGN_TO((uint64_t)ring->ptr, align) - (uint64_t)ring->ptr;
+    size_t pos_addr = (size_t)ring->ptr + ring->pos;
+    size_t alignment_size = ALIGN_TO(pos_addr, align) - pos_addr;
+    size_t pos = (ring->pos + alignment_size + size <= ring->cap) ? ring->pos + alignment_size : ALIGN_TO((size_t)ring->ptr, align) - (size_t)ring->ptr;
     ring->pos = pos + size;
 
     return (char*)ring->ptr + pos;
 }
 
-static inline void* md_ring_allocator_push(md_ring_allocator_t* ring, uint64_t size) {
+static inline void* md_ring_allocator_push(md_ring_allocator_t* ring, size_t size) {
     uint64_t alignment = size <= 2 ? size : MD_RING_ALLOCATOR_DEFAULT_ALIGNMENT;
     return md_ring_allocator_push_aligned(ring, size, alignment);
 }
 
-static inline void md_ring_allocator_pop(md_ring_allocator_t* ring, uint64_t size) {
+static inline void md_ring_allocator_pop(md_ring_allocator_t* ring, size_t size) {
     ASSERT(ring && ring->magic == MD_RING_ALLOCATOR_MAGIC);
     ASSERT(size <= ring->pos);
     ring->pos -= size;
@@ -60,7 +60,7 @@ static inline void md_ring_allocator_reset(md_ring_allocator_t* ring) {
     ring->pos = 0;
 }
 
-static inline void md_ring_allocator_set_pos(md_ring_allocator_t* ring, uint64_t offset) {
+static inline void md_ring_allocator_set_pos(md_ring_allocator_t* ring, size_t offset) {
     ASSERT(ring && ring->magic == MD_RING_ALLOCATOR_MAGIC);
     ASSERT(offset < ring->cap);
     ring->pos = offset;
